AddedTypeAhead

+++ Dtos/NewRentalDto.cs
@@ -8,6 +8,6 @@ namespace VidlyTutorial.Dtos
     public class NewRentalDto
     {
         public int CustomerId { get; set; }
-        public List<int> MoviesIds { get; set; }
+        public List<byte> MoviesIds { get; set; }
     }
 }
